==> notebooks/1-data-loading-processing.ipynb <==
# Data 

## Necessary packages and libraries

import pandas as pd
from sklearn.preprocessing import LabelEncoder
import joblib

## Load Selects cumulative data

# Define the file path
file_path = '../data/selects_cumulative_1971-2019.dta'

# Load the Stata file
data = pd.read_stata(file_path)

# Save the dataframe to a CSV file
csv_file_path = "../data/selects_cumulative_1971-2019.csv"
data.to_csv(csv_file_path, index=False)

print(f"Data saved to {csv_file_path}")

data.head()

## Selecting chosen variables

# Loading original data before subsetting
subset_source = pd.read_csv("../data/selects_cumulative_1971-2019.csv")

# Create a subset with selected columns
selected_columns = ['year', 'userid', 'sex', 'age', 'educ', 'income',
                    'religion','sg1', 'sg9', 'sc1', 'sc7a', 'sc7b',
                    'pi1', 'lr1', 'pm3',
                    'vp1', 'vdn1b', 'pid1', 'pid2b', 'trust1',
                    'weighttot']  
subset = subset_source.loc[:, selected_columns]
subset.head()

## Preprocessing data

# Replace "NaN" with "0" for both numeric and categorical data
subset_clean = subset.fillna(0)

# Further select most important variables
column_drop = ["year", "userid", "lr1", "pid2b", "sc7a", "weighttot", "trust1"]
subset_clean = subset_clean.drop(column_drop, axis=1)

subset_clean.head()

# Dictionary to rename categories
category_rename_mapping = {
    0: 'unknown',  # Example for numeric 0
    'GLP/Vert\'libéraux': "GLP",
    'bdp': 'BDP',
    'centre parties': 'Centre Parties',
    'csp/pcs': 'CSP',
    'cvp/pdc': 'CVP',
    'edu/udf': 'EDU',
    'evp/pep': 'EVP',
    'fdp/prd': 'FDP',
    'fga/avf': 'FGA',
    'fps/psl': 'FPS',
    'gps/pes': 'GPS',
    'ldu/adi': 'LdU',
    'left parties': 'Left Parties',
    'lega': 'Lega',
    'lps/pls': 'LPS',
    'mcg': 'MCG',
    'other comments': 'Other Comments',
    'other parties': 'Other Parties',
    'pda/pdt': 'PdA',
    'poch': 'POCH',
    'psa (psu)': 'PSA',
    'rep. (& vigil.)': 'Rep',
    'right parties': 'Right Parties',
    'sd/ds': 'SD',
    'sol.': 'Sol',
    'sps/pss': 'SP',
    'svp/udc': 'SVP',
    'voted blank': 'Voted Blank'
}

# Convert vdn1b to a categorical text variable
subset_clean['vdn1b'] = subset_clean['vdn1b'].astype('category')

# Rename categories
subset_clean['vdn1b'] = subset_clean['vdn1b'].cat.rename_categories(category_rename_mapping)

# Verify the renaming
print("Renamed categories:", subset_clean['vdn1b'].cat.categories)

# Change LPS to FDP
subset_clean['vdn1b'] = subset_clean['vdn1b'].replace({'LPS': 'FDP'})

subset_clean['vdn1b'].unique()

# Select most important existing parties
subset_selected_parties = subset_clean[subset_clean['vdn1b'].isin(['FDP','CVP','SP','SVP','EVP',
                                                        'GPS', 'GLP'])]

subset_selected_parties

# Identify categorical columns
categorical_cols = subset_selected_parties.select_dtypes(include=['object']).columns
categorical_cols = [col for col in categorical_cols if col != 'vdn1b']

# Encode categorical columns using LabelEncoder
label_encoders = {}
for col in categorical_cols:
    le = LabelEncoder()
    subset_selected_parties[col] = le.fit_transform(subset_selected_parties[col].astype(str))
    label_encoders[col] = le
    
    # Print encoding mapping for each column
    print(f"Encoding mapping for {col}:")
    print(dict(zip(le.classes_, le.transform(le.classes_))))
    print()  # Add a blank line for readability

# Save the encoders for later use
joblib.dump(label_encoders, '../data/models/label_encoders.pkl')

# Verify the conversion of vdn1b
print("vdn1b dtype:", subset_selected_parties['vdn1b'].dtype)  # Should show 'category'

## Save subset with 8 most important parties

# save subset data
subset_selected_parties.to_csv('../data/subset_selected_parties_model.csv', index=False)
